+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/TextInput/TextInputExampleCodeReact.jsx
@@ -17,6 +17,7 @@ function TextInputExampleCodeReact({
       className,
       errorMessage,
       id,
+      isClearable,
       isDisabled,
       isRequired,
       label,
@@ -32,11 +33,12 @@ function TextInputExampleCodeReact({
       <ExampleCodeReactProp displayProp={className ? `className="${className}"` : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={errorMessage ? `errorMessage="${errorMessage}"` : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={id ? `id="${id}"` : null} indentLevel={1} />
+      <ExampleCodeReactProp displayProp={isClearable ? 'isClearable' : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={isDisabled ? 'isDisabled' : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={isRequired ? 'isRequired' : null} indentLevel={1} />
       <ExampleCodeReactProp displayProp={label ? `label="${label}"` : null} indentLevel={1} />
-      <ExampleCodeReactProp displayProp={'onChange={(e) => { /* ... do something w/ e ... */ }'} indentLevel={1} />
-      <ExampleCodeReactProp displayProp={`placeholder="${placeholder}"`} indentLevel={1} />
+      <ExampleCodeReactProp displayProp={isClearable ? 'onClear={(e) => { /* ... do something w/ e ... */ }' : null} indentLevel={1} />
+      <ExampleCodeReactProp displayProp={placeholder && `placeholder="${placeholder}"`} indentLevel={1} />
       <ExampleCodeReactProp displayProp={`value="${value}"`} indentLevel={1} />
       /&gt;
     </>