+++ utah-design-system-website/src/react/components/websiteContent/library/components/forms/TextInput/TextInputExampleProps.jsx
@@ -34,6 +34,7 @@ function TextInputExampleProps({ setState, state }) {
         draftState.props.className = '';
         draftState.props.errorMessage = '';
         draftState.props.id = '';
+        draftState.props.isClearable = false;
         draftState.props.isDisabled = false;
         draftState.props.label = '';
         draftState.props.placeholder = '';
@@ -53,6 +54,7 @@ function TextInputExampleProps({ setState, state }) {
       <TextInput id="props.className" label="Class" className="input--height-small1x" />
       <TextInput id="props.errorMessage" label="Error Message" className="input--height-small1x" />
       <TextInput id="props.id" label="ID" className="input--height-small1x" />
+      <Switch id="props.isClearable" label="Clearable" width={20} />
       <Switch id="props.isDisabled" label="Disabled" width={20} />
       <Switch id="props.isRequired" label="Required" width={20} />
       <TextInput id="props.label" label="Label" className="input--height-small1x" />